Fixed bugs reported by the compiler farm.
git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@7125 a27c51de-2d2b-4d1b-a6e8-f7f87814f453

+++ src/ROSETTA/src/grammar.C
@@ -2350,7 +2350,7 @@ Grammar::buildReferenceToPointerHandlerCode()
                 "struct ReferenceToPointerHandler\n"
                 "   {\n";
 
-     for (int i=0; i < terminalList.size(); i++)
+     for (size_t i=0; i < terminalList.size(); i++)
         {
           s +=  "     virtual void apply(" + terminalList[i]->name + " *&r, const SgName &n, bool traverse) = 0;\n";
         }
@@ -2362,7 +2362,7 @@ Grammar::buildReferenceToPointerHandlerCode()
           "struct ReferenceToPointerHandlerImpl : ReferenceToPointerHandler\n"
           "   {\n";
 
-     for (int i=0; i < terminalList.size(); i++)
+     for (size_t i=0; i < terminalList.size(); i++)
         {
           s +=  "     void apply(" + terminalList[i]->name + " *&r, const SgName &n, bool traverse)\n"
                 "        {\n"
